Remove useless code, fix localisation tests
