--- BTCPayServer/Localization/JsonLocalizationOptions.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.Extensions.Localization;
-
-namespace BTCPayServer.Localization
-{
-    public class JsonLocalizationOptions : LocalizationOptions
-    {
-        public ResourcesType ResourcesType { get; set; } = ResourcesType.TypeBased;
-    }
-}
