+++ BTCPayServer/Localization/JsonLocalizationServiceCollectionExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using BTCPayServer.Localization;
-using BTCPayServer.Localization.Internal;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Localization;
 
@@ -8,35 +7,15 @@ namespace Microsoft.Extensions.DependencyInjection
 {
     public static class JsonLocalizationServiceCollectionExtensions
     {
-        public static IServiceCollection AddJsonLocalization(this IServiceCollection services)
+        public static IServiceCollection AddJsonLocalization(
+           this IServiceCollection services)
         {
             if (services == null)
             {
                 throw new ArgumentNullException(nameof(services));
             }
-
-            services.AddOptions();
 
             AddJsonLocalizationServices(services);
-
-            return services;
-        }
-
-        public static IServiceCollection AddJsonLocalization(
-           this IServiceCollection services,
-           Action<JsonLocalizationOptions> setupAction)
-        {
-            if (services == null)
-            {
-                throw new ArgumentNullException(nameof(services));
-            }
-
-            if (setupAction == null)
-            {
-                throw new ArgumentNullException(nameof(setupAction));
-            }
-            
-            AddJsonLocalizationServices(services, setupAction);
 
             return services;
         }
@@ -44,16 +23,6 @@ namespace Microsoft.Extensions.DependencyInjection
         internal static void AddJsonLocalizationServices(IServiceCollection services)
         {
             services.TryAddSingleton<IStringLocalizerFactory, JsonStringLocalizerFactory>();
-            services.TryAddTransient(typeof(IStringLocalizer<>), typeof(StringLocalizer<>));
-            services.TryAddTransient(typeof(IStringLocalizer), typeof(StringLocalizer));
-        }
-
-        internal static void AddJsonLocalizationServices(
-            IServiceCollection services,
-            Action<JsonLocalizationOptions> setupAction)
-        {
-            AddJsonLocalizationServices(services);
-            services.Configure(setupAction);
         }
     }
 }
