--- BTCPayServer/Localization/ResourcesType.cs
@@ -1,8 +1,0 @@
-﻿namespace BTCPayServer.Localization
-{
-    public enum ResourcesType
-    {
-        CultureBased,
-        TypeBased
-    }
-}
