+++ BTCPayServer/Services/LanguageService.cs
@@ -30,11 +30,6 @@ namespace BTCPayServer.Services
         public LanguageService(IHostingEnvironment environment)
         {
             var path = (environment as HostingEnvironment)?.WebRootPath;
-            if (string.IsNullOrEmpty(path))
-            {
-                //test environment
-                path = Path.Combine(TryGetSolutionDirectoryInfo().FullName,"BTCPayServer", "wwwroot");
-            }
             path = Path.Combine(path, "locales");
             var files = Directory.GetFiles(path, "*.json");
             var result = new List<Language>();
@@ -50,16 +45,6 @@ namespace BTCPayServer.Services
             _languages = result.ToArray();
         }
 
-        public static DirectoryInfo TryGetSolutionDirectoryInfo(string currentPath = null)
-        {
-            var directory = new DirectoryInfo(
-                currentPath ?? Directory.GetCurrentDirectory());
-            while (directory != null && !directory.GetFiles("*.sln").Any())
-            {
-                directory = directory.Parent;
-            }
-            return directory;
-        }
         public Language[] GetLanguages()
         {
             return _languages;
